% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_tuning_plot.R
\name{agg_tuning_plot}
\alias{agg_tuning_plot}
\title{*Generate and save a plot of tuning parameters and the resulting losses.}
\usage{
agg_tuning_plot(tuned_loss, losstype, alg_param, alg_param_opt, output_path,
  execution_id, fold_out_id, algorithm, tuning_method, quiet)
}
\arguments{
\item{tuned_loss}{*(Mean) cross-validation loss (inner tuning routine, i.e. for a given fold).}

\item{losstype}{*Loss function used to tune the individiual learners and the ensemble neg_mse_brier, loglik, auc, accuracy, balanced_accuracy, r2].}

\item{alg_param}{*Tuning parameters [list - misc].}

\item{alg_param_opt}{*Index of chosen tuning parameters set.}

\item{output_path}{*Directory within which all model output (e.g. tuning graphs) is stored [path].}

\item{execution_id}{*Unique ID generated at the beginning of each model construction process - all output is associated with this ID [character].}

\item{fold_out_id}{*Fold on which to predict [integer].}

\item{algorithm}{*Name of individual algorithm to be tuned [character].}

\item{tuning_method}{*Tuning method [character].}
}
\value{
*.
}
\description{
*Aggregate rather than parameter-specific tuning plot.
}

