% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg.R
\name{alg}
\alias{alg}
\title{*Combine an algorithm with a list of tuning parameters.}
\usage{
alg(algorithm)
}
\arguments{
\item{algorithm}{*Name of individual algorithm to be tuned [character].}
}
\value{
*List with the algorithm [algorithm] and the parameters [parameter].
}
\description{
*.
}

