% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_stratification.R
\name{blocked_stratification}
\alias{blocked_stratification}
\title{*Split a data.table into stratified cluster-level subsets.}
\usage{
blocked_stratification(df, stratify_var, cluster_var, prop_vector, stratify_cut)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{stratify_var}{*Name of variable on which to stratify (when generating CV folds, test vs. holdout splits)  [character].}

\item{cluster_var}{*Name of variable containing observation level cluster identifiers (integer) [character].}

\item{prop_vector}{*Number of intervals into which the stratification variable is split (if fold_method==stratification) [integer] .}

\item{stratify_cut}{*Number of intervals into which the stratification variable is split (if fold_method==stratification) [integer] .}
}
\value{
*Vector of subset IDs (to be merged into original data.table).
}
\description{
*.
}

