% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_data.R
\name{complete_data}
\alias{complete_data}
\title{*Convert a data.table into a data.table containing only complete observations (and potentially featuring missigness indicators).}
\usage{
complete_data(df, missing_var, non_missing_var, assignment_var,
  missing_mode = ExpLearning.Default$missing_mode,
  holdout_var = ExpLearning.Default$holdout_var,
  quiet = ExpLearning.Default$quiet)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{missing_var}{*Names of all RHS variables that need to exist for every observation in the datatable used for model construction [character].}

\item{non_missing_var}{*Names of all RHS variables that do not need need to exist for every observation in the datatable used for model construction [character].}

\item{assignment_var}{*Name of outcome variable [character].}

\item{missing_mode}{*How to handle missingness - (a) omit all missing observations (missing_omit), (b) contruct missingness indicators (and median impute the missing values) treating the missingness indicators as balanced covariates (missing_indic_balanced) or (c) contruct missingness indicators (and median impute the missing values) treating the missingness indicators as unbalanced covariates (missing_indic_unbalanced) [character].}

\item{quiet}{*Verbosity settings [0 (verbose), 1 (print only key stats and progress updates), 2 (silent) - 0.5 (test and development mode)].}
}
\value{
*Complete data.table (df) alongside a list of variables affected by missingness (miss_var).
}
\description{
*.
}

