% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_split_grid.R
\name{expand_split_grid}
\alias{expand_split_grid}
\title{*Convert a list of tuning parameters into a list of data.tables each containing one tuning parameter combination.}
\usage{
expand_split_grid(alg_param_tuning, alg_param_default = NULL, iter = NULL,
  param_split_method)
}
\arguments{
\item{alg_param_tuning}{*Tuning parameters - tuning [list - misc].}

\item{alg_param_default}{*Tuning parameters - default [list - misc].}

\item{param_split_method}{*Method used to split and combine a parameter list into a (tuning) parameter grid [character].}
}
\value{
*Formatted parameter tuning grid.
}
\description{
*.
}

