% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpLearning.Default_data.R
\docType{data}
\name{ExpLearning.Default}
\alias{ExpLearning.Default}
\title{*[Dataset] Default values for all parameters associated with the ExpLearning Package.}
\format{An object of class \code{list} of length 81.}
\usage{
data(ExpLearning.Default)
}
\description{
*.
}
\keyword{datasets}

