% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpLearning.MLParameter_data.R
\docType{data}
\name{ExpLearning.MLParameter}
\alias{ExpLearning.MLParameter}
\title{*[Dataset] Overview of all ml parameters, i.e. tuning parameters, associated with the learners in the ExpLearning Package.}
\format{An object of class \code{list} of length 35.}
\usage{
data(ExpLearning.MLParameter)
}
\description{
*.
}
\keyword{datasets}

