% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpLearning.Parameter_data.R
\docType{data}
\name{ExpLearning.Parameter}
\alias{ExpLearning.Parameter}
\title{*[Dataset] Overview of all parameters associated with the ExpLearning Package.}
\format{An object of class \code{list} of length 81.}
\usage{
data(ExpLearning.Parameter)
}
\description{
*.
}
\keyword{datasets}

