% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param.R
\name{extract_param}
\alias{extract_param}
\title{*Convert tuning parameters returned as part of the results into a list of parameters which can be passed to the next round as a custom_param argument.}
\usage{
extract_param(model_result)
}
\arguments{
\item{model_result}{*List of results returned by the effect_on_group function [list-misc].}
}
\value{
*Formatted list of tuning parameters.
}
\description{
*.
}

