% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output}
\alias{format_output}
\title{*Format the model output.}
\usage{
format_output(model_result_raw, setting, dict)
}
\arguments{
\item{model_result_raw}{*List of outputs returned from the holdout_(opt)_outcome_predict or kfold_(opt)_outcome_predict functions [list - misc].}

\item{setting}{*List with key settings generated at the beginning of each model construction process [list - misc].}

\item{dict}{*List with the names of all key variables initialised at the beginning of each model construction process [list - character].}
}
\value{
*A list with the formatted model output.
}
\description{
*.
}

