% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fold.R
\name{gen_fold}
\alias{gen_fold}
\title{*Generate cluster-level CV folds.}
\usage{
gen_fold(df, cluster_var = ExpLearning.Default$cluster_var,
  nfold = ExpLearning.Default$holdout_var,
  fold_method = ExpLearning.Default$fold_method, stratify_var = NULL,
  stratify_cut = ExpLearning.Default$stratify_cut,
  fold_var = ExpLearning.Default$fold_var,
  return_index = ExpLearning.Default$return_index,
  weight_var = ExpLearning.Default$weight_var,
  quiet = ExpLearning.Default$quiet)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{cluster_var}{*Name of variable containing observation level cluster identifiers (integer) [character].}

\item{nfold}{*Number of CV folds (used in both the inner and outer CV routine) [integer].}

\item{fold_method}{*Method used to generate CV folds (and the train-holdout split) [cluster_count (target cluster count), obs_count (target observation count), stratification (target stratification)].}

\item{stratify_var}{*Name of variable on which to stratify (when generating CV folds, test vs. holdout splits)  [character].}

\item{stratify_cut}{*Number of intervals into which the stratification variable is split (if fold_method==stratification) [integer] .}

\item{fold_var}{*Name of variable containing observation level fold identifiers (integer) [character].}

\item{return_index}{*Whether to return a modified data.table with an appended holdout, fold column [FALSE] or a vector of holdout, fold ids [logical - TRUE, FALSE].}
}
\value{
*Modified data.table containing fold IDs for each observation.
}
\description{
*Holdout set construction optionally targeting (i) the cluster count (ii) the sample size or (iii) stratification.
}
\examples{
gen_fold(test_df, nfold=5, cluster_var="cluster", fold_method="cluster_count", 
 stratify_var="treatment", stratify_cut=20) 
gen_fold(test_df, nfold=5, cluster_var="cluster", fold_method="obs_count", 
 stratify_var="treatment", stratify_cut=20) 
gen_fold(test_df, nfold=5, cluster_var="cluster", fold_method="stratification", 
 stratify_var="treatment", stratify_cut=20) 
}

