% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_custom.R
\name{glmnet_custom}
\alias{glmnet_custom}
\title{.}
\usage{
glmnet_custom(x, y, family = c("gaussian", "binomial", "poisson",
  "multinomial", "cox", "mgaussian"), weights, offset = NULL, alpha = 1,
  nlambda = 100, lambda.min.ratio = ifelse(nobs < nvars, 0.01, 0.0001),
  lambda = NULL, standardize = TRUE, intercept = TRUE,
  thresh = 0.0000001, dfmax = nvars + 1, pmax = min(dfmax * 2 + 20,
  nvars), exclude, penalty.factor = rep(1, nvars), lower.limits = -Inf,
  upper.limits = Inf, maxit = 100000, type.gaussian = ifelse(nvars < 500,
  "covariance", "naive"), type.logistic = c("Newton", "modified.Newton"),
  standardize.response = FALSE, type.multinomial = c("ungrouped",
  "grouped"))
}
\description{
.
}

