% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout_analysis.R
\name{holdout_analysis}
\alias{holdout_analysis}
\title{*Analyse the predictive performance of the out of sample predictions derived from the model.}
\usage{
holdout_analysis(model_result_raw, dict, setting)
}
\arguments{
\item{model_result_raw}{*List of outputs returned from the holdout_(opt)_outcome_predict or kfold_(opt)_outcome_predict functions [list - misc].}

\item{dict}{*List with the names of all key variables initialised at the beginning of each model construction process [list - character].}

\item{setting}{*List with key settings generated at the beginning of each model construction process [list - misc].}
}
\value{
*List of losses, a p-value and a data.table with the original data alongside the final, out of sample predictions for the ensemble and each of the individual learners.
}
\description{
*.
}

