% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout_outcome_predict.R
\name{holdout_outcome_predict}
\alias{holdout_outcome_predict}
\title{*Construct an ensemble model using a holdout set and generate out of sample predictions based on this model [Interface].}
\usage{
holdout_outcome_predict(df, dict, setting)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{dict}{*List with the names of all key variables initialised at the beginning of each model construction process [list - character].}

\item{setting}{*List with key settings generated at the beginning of each model construction process [list - misc].}
}
\value{
*A list with out of sample predictions as well as hyperparameters and ensemble weights.
}
\description{
*.
}

