% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout_predict.R
\name{holdout_predict}
\alias{holdout_predict}
\title{*Construct an ensemble model using a training subset of the data and generate out of sample predictions based on this model.}
\usage{
holdout_predict(df, LHS, RHS, RHS_ext,
  custom_param = ExpLearning.Default$custom_param,
  cluster_var = ExpLearning.Default$cluster_var,
  prediction_var = ExpLearning.Default$prediction_var,
  ensemble_agg = ExpLearning.Default$ensemble_agg,
  weight_var = ExpLearning.Default$weight_var,
  id_var = ExpLearning.Default$id_var,
  tuning_method = ExpLearning.Default$tuning_method,
  fold_method = ExpLearning.Default$fold_method,
  rand_grid_iter = ExpLearning.Default$rand_grid_iter,
  losstype = ExpLearning.Default$losstype, task = ExpLearning.Default$task,
  max_core = ExpLearning.Default$max_core,
  holdout_var = ExpLearning.Default$holdout_var,
  execution_id = eval(parse(text = ExpLearning.Default$exceution_id)),
  predictor = ExpLearning.Default$predictor,
  innerfold = ExpLearning.Default$innerfold,
  insample = ExpLearning.Default$insample,
  quiet = ExpLearning.Default$quiet, output_path = eval(parse(text =
  ExpLearning.Default$output_path)), fold_var = ExpLearning.Default$fold_var,
  round_digit = ExpLearning.Default$round_digit)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{LHS}{*Name of the independent variable [character].}

\item{RHS}{*Names of all dependent variables (may incl. group_var, control_var and id_var) [character].}

\item{RHS_ext}{*Names of all dependent variables (may incl. group_var, control_var and id_var) [character].}

\item{custom_param}{*Custom tuning parameters supplied to the prediction function and used in place of the default values (if tuning_method==""no_tuning"") [list - misc].}

\item{cluster_var}{*Name of variable containing observation level cluster identifiers (integer) [character].}

\item{ensemble_agg}{*Method used to aggregate the individual learners into an ensemble learner [ols  (linear regression on individual learners predictions in the case of a regression task vs. logit regression on individual learner predictions in the case of a classification task), nnls (weights derived through non-negative least squares)]. [character].}

\item{tuning_method}{*Tuning method [character].}

\item{fold_method}{*Method used to generate CV folds (and the train-holdout split) [cluster_count (target cluster count), obs_count (target observation count), stratification (target stratification)].}

\item{rand_grid_iter}{*Number of random grid search iterations (performed if tuning_method==""rand_grid"") [integer].}

\item{losstype}{*Loss function used to tune the individiual learners and the ensemble neg_mse_brier, loglik, auc, accuracy, balanced_accuracy, r2].}

\item{task}{*The nature of the prediction task  [regression (regression), classification (classification)].}

\item{max_core}{*The maximum number of cores to be used in all parallelised loops [integer].}

\item{execution_id}{*Unique ID generated at the beginning of each model construction process - all output is associated with this ID [character].}

\item{predictor}{*List of algorithms included in ensemble learner [ols (linear regression), rf (random forest (classification OR regression)), avg (simple mean), elnet (elastic net), logit (logistic regression), tree (tree (classification or regression)), xgb (gradient boosted tree (classification or regression))] [character].}

\item{innerfold}{*Number of CV folds used in inner CV routine [integer].}

\item{insample}{*Whether (with holdout==TRUE) the holdout  set predictions are based on the whole sample (TRUE) or only the holdoutset (FALSE) [logical - TRUE, FALSE].}

\item{quiet}{*Verbosity settings [0 (verbose), 1 (print only key stats and progress updates), 2 (silent) - 0.5 (test and development mode)].}

\item{output_path}{*Directory within which all model output (e.g. tuning graphs) is stored [path].}

\item{fold_var}{*Name of variable containing observation level fold identifiers (integer) [character].}

\item{holdout}{*Whether or not to construct and use a holdout set [logical - TRUE, FALSE].}

\item{prediction}{*Vector of predictions [numeric].}

\item{weight}{*Vector of observation weights [integer or numeric].}
}
\value{
*A list with out of sample predictions as well as hyperparameters and ensemble weights.
}
\description{
*.
}

