% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_analysis.R
\name{kfold_analysis}
\alias{kfold_analysis}
\title{*Analyse the predictive performance of the CV-based (out of sample) predictions derived from the model.}
\usage{
kfold_analysis(model_result_raw, perm_return, dict, setting)
}
\arguments{
\item{model_result_raw}{*List of outputs returned from the holdout_(opt)_outcome_predict or kfold_(opt)_outcome_predict functions [list - misc].}

\item{perm_return}{*List of outputs returned from the kfold_(opt)_outcome_predict functions for each of the permuted data sets [list - misc].}

\item{dict}{*List with the names of all key variables initialised at the beginning of each model construction process [list - character].}

\item{setting}{*List with key settings generated at the beginning of each model construction process [list - misc].}
}
\value{
*List of losses, a p-value and a data.table with the original data alongside the final, CV-based (out of sample predictions) for the ensemble and each of the individual learners.
}
\description{
*.
}

