% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_outcome_predict.R
\name{kfold_outcome_predict}
\alias{kfold_outcome_predict}
\title{*Construct an ensemble model using the entire data and generate CV-based (out of sample) predictions based on this model [Interface].}
\usage{
kfold_outcome_predict(df, dict, setting, permute, perm_test = NULL,
  quiet = ExpLearning.Default$quiet)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{dict}{*List with the names of all key variables initialised at the beginning of each model construction process [list - character].}

\item{setting}{*List with key settings generated at the beginning of each model construction process [list - misc].}

\item{permute}{*Whther to construct the ensemble using the original data or a permuted version [logical].}

\item{perm_test}{*(Permutation) test(s) to perform [character].}

\item{quiet}{*Verbosity settings [0 (verbose), 1 (print only key stats and progress updates), 2 (silent) - 0.5 (test and development mode)].}
}
\value{
*A list with out of sample predictions as well as hyperparameters and ensemble weights.
}
\description{
*
}

