% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitalg.R
\name{logitalg}
\alias{logitalg}
\title{*Fit a logistic regression to a subset of the data and return predictions for (i) a (different) subset of the data and (ii) the original data.}
\usage{
logitalg(df, LHS, RHS, from, to, weight_var = ExpLearning.Default$weight_var,
  task = ExpLearning.Default$task,
  alg_param = ExpLearning.MLParameter[["logit_param"]][[1]],
  quiet = ExpLearning.Default$quiet, ...)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{LHS}{*Name of the independent variable [character].}

\item{RHS}{*Names of all dependent variables (may incl. group_var, control_var and id_var) [character].}

\item{from}{*Observations on which to train an algorithm [integer].}

\item{to}{*Observations on which to predict [integer].}

\item{task}{*The nature of the prediction task  [regression (regression), classification (classification)].}

\item{alg_param}{*Tuning parameters [list - misc].}

\item{quiet}{*Verbosity settings [0 (verbose), 1 (print only key stats and progress updates), 2 (silent) - 0.5 (test and development mode)].}

\item{weight}{*Vector of observation weights [integer or numeric].}
}
\value{
*Vector of predictions (predicted probability of Y==1) [pred pred_is] and the fitted learner [learner].
}
\description{
*Logistic regression constructed using glm (..., family=binomial)  [base R].  Observation weights (weight_var) are handled internally (method: minimise weighted loss function) in the training stage (not taken into account at the prediction stage).
}

