% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict_function_template.R
\name{model_predict_function_template}
\alias{model_predict_function_template}
\title{*Ensemble prediction function template.}
\usage{
model_predict_function_template(...)
}
\description{
*.
}

