% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.R
\name{perm}
\alias{perm}
\title{*Perform cluster-level permutations of a specified variable in a given data.table.}
\usage{
perm(df, assignment_var, control_var, control_var_unbalanced,
  cluster_var = ExpLearning.Default$cluster_var,
  fold_var = ExpLearning.Default$fold_var,
  within_fold = ExpLearning.Default$within_fold,
  perm_strategy = ExpLearning.Default$perm_strategy,
  perm_test = ExpLearning.Default$perm_test)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{assignment_var}{*Name of outcome variable [character].}

\item{control_var}{*Vector of control variable names (balanced across treatment conditions ) (RHS) [character].}

\item{control_var_unbalanced}{*Vector of control variable names ((potentially) unabalanced across treatment conditions, i.e. random assignemnt is assumed to hold conditional on these variables) (RHS) [character].}

\item{cluster_var}{*Name of variable containing observation level cluster identifiers (integer) [character].}

\item{fold_var}{*Name of variable containing observation level fold identifiers (integer) [character].}

\item{within_fold}{*Whether  permutations are performed within folds or across folds [logical - TRUE, FALSE].}

\item{perm_strategy}{*Whether to permute the assignment variable by (a) exchanging variable values at the cluster level (entire_cluster) or (b) exchanging lvariable values within clusters (within_cluster) [character].}

\item{perm_test}{*(Permutation) test(s) to perform [character].}
}
\value{
*Modified data.table with permuted variable.
}
\description{
*.
}

