% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_analysis.R
\name{perm_analysis}
\alias{perm_analysis}
\title{*Generate a permutation testing plot (plot of the permuted losses and the model loss).}
\usage{
perm_analysis(main_loss, perm_loss, pval, pval_95_ci, perm_test,
  losstype = ExpLearning.Default$losstype)
}
\arguments{
\item{main_loss}{*Out of sample ensemble loss [numeric].}

\item{perm_loss}{*Out of sample permutation losses [numeric].}

\item{pval}{*Final p-value.}

\item{losstype}{*Loss function used to tune the individiual learners and the ensemble neg_mse_brier, loglik, auc, accuracy, balanced_accuracy, r2].}
}
\value{
*Permutation testing plot.
}
\description{
*.
}

