% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_testing.R
\name{perm_testing}
\alias{perm_testing}
\title{*Derive a permutation testing based p-value for a given model.}
\usage{
perm_testing(df, assignment_var, control_var, control_var_unbalanced,
  perm_df_seq = NULL, model_predict_function = NULL,
  perm_testing_mode = ExpLearning.Default$perm_testing_mode,
  prediction_var = ExpLearning.Default$prediction_var,
  permutation = ExpLearning.Default$permutation,
  cluster_var = ExpLearning.Default$cluster_var,
  fold_var = ExpLearning.Default$fold_var,
  losstype = ExpLearning.Default$losstype,
  weight_var = ExpLearning.Default$weight_var,
  within_fold = ExpLearning.Default$within_fold,
  max_core = ExpLearning.Default$max_core,
  perm_strategy = ExpLearning.Default$perm_strategy,
  task = ExpLearning.Default$task,
  perm_test = ExpLearning.Default$perm_test,
  quiet = ExpLearning.Default$quiet)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{assignment_var}{*Name of outcome variable [character].}

\item{control_var}{*Vector of control variable names (balanced across treatment conditions ) (RHS) [character].}

\item{control_var_unbalanced}{*Vector of control variable names ((potentially) unabalanced across treatment conditions, i.e. random assignemnt is assumed to hold conditional on these variables) (RHS) [character].}

\item{model_predict_function}{*Ensemble predict function.}

\item{perm_testing_mode}{*Wheter to permute the data or to (re)use a provided, (permuted) df as the basis for the permutationt test [character].}

\item{prediction_var}{*Name of variable containing observation level predictions (numeric) [character].}

\item{permutation}{*Number of permutations performed to generate the p-values [integer].}

\item{cluster_var}{*Name of variable containing observation level cluster identifiers (integer) [character].}

\item{fold_var}{*Name of variable containing observation level fold identifiers (integer) [character].}

\item{losstype}{*Loss function used to tune the individiual learners and the ensemble neg_mse_brier, loglik, auc, accuracy, balanced_accuracy, r2].}

\item{weight_var}{*Name of variable containing observation level weights (integer) [character].}

\item{within_fold}{*Whether  permutations are performed within folds or across folds [logical - TRUE, FALSE].}

\item{max_core}{*The maximum number of cores to be used in all parallelised loops [integer].}

\item{perm_strategy}{*Whether to permute the assignment variable by (a) exchanging variable values at the cluster level (entire_cluster) or (b) exchanging lvariable values within clusters (within_cluster) [character].}

\item{task}{*The nature of the prediction task  [regression (regression), classification (classification)].}

\item{perm_test}{*(Permutation) test(s) to perform [character].}

\item{quiet}{*Verbosity settings [0 (verbose), 1 (print only key stats and progress updates), 2 (silent) - 0.5 (test and development mode)].}
}
\value{
*List with p-values, permuted losses and model losses.
}
\description{
*.
}

