% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{*Prepare a data.table containing the raw data for further processing with the ExpLearning Package and generate a configuration dictionary.}
\usage{
prepare_data(df, assignment_var, group_var, fold_var = NULL,
  holdout_var = NULL, control_var = NULL, weight_var = NULL,
  cluster_var = NULL, id_var = NULL, control_var_unbalanced = NULL,
  clusterweight_norm = ExpLearning.Default$clusterweight_norm,
  task = ExpLearning.Default$task, quiet = ExpLearning.Default$quiet)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{assignment_var}{*Name of outcome variable [character].}

\item{group_var}{*Name of predictor variable [character].}

\item{fold_var}{*Name of variable containing observation level fold identifiers (integer) [character].}

\item{control_var}{*Vector of control variable names (balanced across treatment conditions ) (RHS) [character].}

\item{cluster_var}{*Name of variable containing observation level cluster identifiers (integer) [character].}

\item{id_var}{*Name of variable containing unique observation level identifiers (integer) [character].}

\item{control_var_unbalanced}{*Vector of control variable names ((potentially) unabalanced across treatment conditions, i.e. random assignemnt is assumed to hold conditional on these variables) (RHS) [character].}

\item{clusterweight_norm}{*Whether to normalise weights at the cluster level [logical - TRUE, FALSE].}

\item{task}{*The nature of the prediction task  [regression (regression), classification (classification)].}

\item{quiet}{*Verbosity settings [0 (verbose), 1 (print only key stats and progress updates), 2 (silent) - 0.5 (test and development mode)].}

\item{weight}{*Vector of observation weights [integer or numeric].}
}
\value{
*List with the modified data.table (df) and the configuration dictionary (dict).
}
\description{
*.
}

