% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfalg.R
\name{rfalg}
\alias{rfalg}
\title{*Fit a random forest model (classification regression) to a subset of the data and return predictions for (i) a (different) subset of the data and (ii) the original data.}
\usage{
rfalg(df, LHS, RHS, from, to, weight_var = ExpLearning.Default$weight_var,
  task = ExpLearning.Default$task,
  alg_param = ExpLearning.MLParameter[["rf_param"]][[1]],
  quiet = ExpLearning.Default$quiet, ...)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{LHS}{*Name of the independent variable [character].}

\item{RHS}{*Names of all dependent variables (may incl. group_var, control_var and id_var) [character].}

\item{from}{*Observations on which to train an algorithm [integer].}

\item{to}{*Observations on which to predict [integer].}

\item{task}{*The nature of the prediction task  [regression (regression), classification (classification)].}

\item{alg_param}{*Tuning parameters [list - misc].}

\item{quiet}{*Verbosity settings [0 (verbose), 1 (print only key stats and progress updates), 2 (silent) - 0.5 (test and development mode)].}

\item{weight}{*Vector of observation weights [integer or numeric].}
}
\value{
*Vector of predictions (predicted value of Y  OR  predicted probability of Y==1) [pred pred_is] and the fitted learner [learner].
}
\description{
*Random forest constructed using ranger [ranger]. Observation weights (weight_var) are handled internally (method: Observation with heigher weights are selected with higher probability in the bootstrap samples used to construct the individual trees) in the training stage (not taken into account at the prediction stage).
}

