% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_df_data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{*[Dataset] Simple data.table useful for testing purposes.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 10000 rows and 13 columns.}
\usage{
data(test_df)
}
\description{
*.
}
\keyword{datasets}

