% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_grid.R
\name{tuning_grid}
\alias{tuning_grid}
\title{.}
\usage{
tuning_grid(df, LHS, RHS, algorithm, alg_param_grid,
  weight_var = ExpLearning.Default$weight_var,
  task = ExpLearning.Default$task, max_core = ExpLearning.Default$max_core,
  losstype = ExpLearning.Default$losstype,
  quiet = ExpLearning.Default$quiet)
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{LHS}{*Name of the independent variable [character].}

\item{RHS}{*Names of all dependent variables (may incl. group_var, control_var and id_var) [character].}

\item{algorithm}{*Name of individual algorithm to be tuned [character].}

\item{alg_param_grid}{}

\item{task}{*The nature of the prediction task  [regression (regression), classification (classification)].}

\item{max_core}{*The maximum number of cores to be used in all parallelised loops [integer].}

\item{losstype}{*Loss function used to tune the individiual learners and the ensemble neg_mse_brier, loglik, auc, accuracy, balanced_accuracy, r2].}

\item{quiet}{*Verbosity settings [0 (verbose), 1 (print only key stats and progress updates), 2 (silent) - 0.5 (test and development mode)].}
}
\value{
.
}
\description{
.
}

