% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_ensemble.R
\name{wald_ensemble}
\alias{wald_ensemble}
\title{*Perform a series of Wald tests to test the significance of the group_var and control_var in a regression of the assignment var on the  group_var, control_var and unbalanced_control_var.}
\usage{
wald_ensemble(df, assignment_var, group_var, control_var,
  control_var_unbalanced, wald_test = ExpLearning.Default$wald_test,
  holdout_var = ExpLearning.Default$holdout_var,
  permutation = ExpLearning.Default$permutation,
  perm_strategy = ExpLearning.Default$perm_strategy,
  max_core = ExpLearning.Default$max_core,
  cluster_var = ExpLearning.Default$cluster_var,
  quiet = ExpLearning.Default$quiet, perm_test = "baseline_signal")
}
\arguments{
\item{df}{*Datatable containing the data [data.table].}

\item{assignment_var}{*Name of outcome variable [character].}

\item{group_var}{*Name of predictor variable [character].}

\item{control_var}{*Vector of control variable names (balanced acroos treatment conditions ) (RHS) [character].}

\item{control_var_unbalanced}{*Vector of control variable names ((potentially) unabalanced acroos treatment conditions, i.e. random assignemnt is assumed to hold conditional on these variables) (RHS) [character].}

\item{wald_test}{*Whether to conduct and report a series of Wald tests based on a linear model of the LHS as a function of all RHS variables (note: only applicable in the case of the holdout, optimised holdout, optimized CV approach) [logical].}

\item{holdout_var}{*Name of variable containing observation level training vs. holdout set identifiers (TRUE, FALSE - TRUE: holdout) [character].}

\item{permutation}{*Number of permutations performed to generate the p-values [integer].}

\item{perm_strategy}{*Whether to permute the assignment variable by (a) exchanging variable values at the cluster level (entire_cluster) or (b) exchanging lvariable values within clusters (within_cluster) [character].}

\item{max_core}{*The maximum number of cores to be used in all parallelised loops [integer].}

\item{cluster_var}{*Name of variable containing observation level cluster identifiers (integer) [character].}

\item{quiet}{*Verbosity settings [0 (verbose), 1 (print only key stats and progress updates), 2 (silent) - 0.5 (test and development mode)].}

\item{perm_test}{*(Permutation) test(s) to perform [character].}
}
\value{
*A list of p-values for each of the wald tests.
}
\description{
*.
}

